requirements.txt
#_libgcc_mutex==0.1
#_openmp_mutex==5.1
absl-py==1.4.0
accelerate==0.23.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
array-record==0.4.1
asttokens==2.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
babel==2.12.1
#backcall==0.2.0
backoff==2.2.1
backoff==2.2.1
blessed==1.20.0
#bleurt==0.0.2
#ca-certificates==2023.08.22
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
chex==0.1.83
click==8.1.7
clu==0.0.9
cmake==3.27.5
colorama==0.4.6
comm==0.1.4
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.11.0
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
dm-tree==0.1.8
docstring-parser==0.15
editdistance==0.6.2
einops==0.7.0
enum34==1.1.10
etils==1.5.0
eval-type-backport==0.1.3
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.103.1
ffmpy==0.3.2
filelock==3.12.4
flatbuffers==23.5.26
flax==0.7.4
fonttools==4.42.1
frozenlist==1.4.0
fschat==0.2.28
fsspec==2023.6.0
func-timeout==4.3.5
future==0.18.3
gast==0.4.0
gin-config==0.5.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
gpustat==1.1.1
gradio==4.19.2
gradio-client==0.10.1
gradio-highlightedtextbox==0.0.11
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.21.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.2
ipython==8.16.1
jax==0.4.16
jaxlib==0.4.16
jedi==0.19.1
jinja2==3.1.2
joblib==1.3.2
jsonlines==4.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==8.4.0
jupyter-core==5.4.0
keras==2.13.1
kiwisolver==1.4.5
latex2mathml==3.77.0
#ld_impl_linux-64==2.38
#libclang==16.0.6
#libffi==3.4.4
#libgcc-ng==11.2.0
#libgomp==11.2.0
#libstdcxx-ng==11.2.0
lit==16.0.6
lxml==4.9.3
markdown==3.4.4
markdown-it-py==3.0.0
markdown2==2.4.10
markupsafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdtex2html==1.3.0
mdurl==0.1.2
mesh-tensorflow==0.1.21
ml-collections==0.1.1
ml-dtypes==0.3.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.15
#ncurses==6.4
nest-asyncio==1.5.8
networkx==3.1
nh3==0.2.14
nltk==3.8.1
numpy==1.24.3
# nvidia-cublas-cu11==11.10.3.66
# nvidia-cublas-cu12==12.3.2.9
# nvidia-cuda-cupti-cu11==11.7.101
# nvidia-cuda-nvrtc-cu11==11.7.99
# nvidia-cuda-nvrtc-cu12==12.3.52
# nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cuda-runtime-cu12==12.3.52
# nvidia-cudnn-cu11==8.5.0.96
# nvidia-cudnn-cu12==8.9.6.50
# nvidia-cufft-cu11==10.9.0.58
# nvidia-curand-cu11==10.2.10.91
# nvidia-cusolver-cu11==11.4.0.1
# nvidia-cusparse-cu11==11.7.4.91
# nvidia-ml-py==12.535.108
# nvidia-nccl-cu11==2.14.3
# nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openai==0.28.0
#openssl==3.0.10
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.3.5
orjson==3.9.15
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.0.1
pip==23.2.1
platformdirs==3.11.0
portalocker==2.8.2
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic-core==2.16.3
pydub==0.25.1
pyglove==0.4.3
pygments==2.16.1
pyparsing==3.1.1
#python==3.9.18
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2023.3.post1
pyyaml==6.0.1
pyzmq==25.1.1
#readline==8.2
referencing==0.33.0
regex==2023.8.8
#repe==0.1
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.5.3
rouge-score==0.1.2
rpds-py==0.18.0
rsa==4.9
ruff==0.3.0
sacrebleu==2.3.1
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
semantic-version==2.10.0
sentencepiece==0.1.99
seqio==0.0.18
setuptools==68.0.0
shellingham==1.5.4
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
#sqlite==3.41.2
stack-data==0.6.3
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
t5==0.9.3
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.13.0
tensorflow-hub==0.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
tensorflow-text==2.13.0
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
#tensorrt-libs==8.6.1
tensorstore==0.1.43
termcolor==2.3.0
tf-slim==1.1.0
tfds-nightly==4.9.3.dev202309200044
threadpoolctl==3.2.0
tiktoken==0.5.1
#tk==8.6.12
tokenizers==0.14.1
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.0
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.35.0
triton==2.0.0
#truthfulqa==0.0.1
tsv==1.2
typer==0.9.0
typing-extensions==4.10.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
wavedrom==2.0.3.post3
wcwidth==0.2.6
websockets==11.0.3
werkzeug==2.3.7
wheel==0.38.4
wrapt==1.15.0
xxhash==3.3.0
#xz==5.4.2
yarl==1.9.2
zipp==3.17.0
#zlib==1.2.13
